requirements.txt
tweepy
python-dotenv
azure-storage-blob
